+++ src/application/application_api.rs
@@ -6,7 +6,12 @@ use crate::{query_db::search::Search, scraper::item_result::ItemResult};
 
 #[async_trait]
 pub trait ApplicationApi {
-    async fn add_search(&mut self, name: &str, query: &str) -> Result<(), Box<dyn Error>>;
+    async fn add_search(
+        &mut self,
+        name: &str,
+        query: &str,
+        max_price: Option<i32>,
+    ) -> Result<(), Box<dyn Error>>;
     async fn delete_search(&mut self, name: &str) -> Result<(), Box<dyn Error>>;
     async fn list(&self) -> Result<Vec<Arc<Search>>, Box<dyn Error>>;
     async fn scrape(&self, notify: Option<bool>) -> Result<Vec<Arc<ItemResult>>, Box<dyn Error>>;